# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sahara Triage Voice Agent API",
    "summary": "Conversational health-triage voice agent: Supabase-backed triage, scheduling, and clinician workflows.",
    "description": "Patient signup and voice triage (browser WebSocket or Africa's Talking phone channel), automatic doctor-slot routing, clinician queues with embedded triage reports, prescriptions, and consent-gated call workflows.\n\nAuth: Supabase session JWT as `Authorization: Bearer <token>` (use the Authorize button). The browser voice channel is a WebSocket and is documented in the README message catalog; phone callbacks accept `application/x-www-form-urlencoded` and return TwiML XML.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api.sahara.example.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/api/voice/webhook": {
      "post": {
        "tags": [
          "voice"
        ],
        "summary": "Africa's Talking inbound-call callback",
        "description": "Accepts `application/x-www-form-urlencoded` (callerNumber, sessionId, isActive) and returns TwiML XML (`<Say>`, `<GetDigits>`, `<Record>`, `<Dial>`).",
        "operationId": "voice_webhook_api_voice_webhook_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "403": {
            "description": "Invalid webhook secret"
          }
        }
      }
    },
    "/api/voice/consent": {
      "post": {
        "tags": [
          "voice"
        ],
        "summary": "Recording-consent IVR callback",
        "description": "Accepts form-data (`conversation_id`, `dtmfDigits`) and returns TwiML XML. Digit 1 starts recorded triage turns; anything else escalates without recording.",
        "operationId": "voice_consent_api_voice_consent_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "403": {
            "description": "Invalid webhook secret"
          }
        }
      }
    },
    "/api/voice/recording": {
      "post": {
        "tags": [
          "voice"
        ],
        "summary": "One recorded triage turn",
        "description": "Accepts form-data (`conversation_id`, `recordingUrl`, `recording_consent`), runs STT \u2192 safety/triage \u2192 auto-routing, and returns the next TwiML XML prompt including the appointment time once triage is reached.",
        "operationId": "voice_recording_api_voice_recording_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "403": {
            "description": "Invalid webhook secret"
          }
        }
      }
    },
...
    "/api/appointments/route": {
      "post": {
        "tags": [
          "appointments"
        ],
        "summary": "Auto-book earliest free doctor slot",
        "description": "Book the earliest free doctor slot for a triaged conversation.\n\nCalled by the web-voice client after TRIAGE_UPDATE; the phone channel\ncalls the same function internally in voice_webhook.py.",
        "operationId": "route_my_appointment_api_appointments_route_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteAppointmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteAppointmentResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AppointmentOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "patient_id": {
            "type": "string",
            "format": "uuid",
            "title": "Patient Id"
          },
          "conversation_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "tier": {
            "type": "string",
            "title": "Tier",
            "examples": [
              "AMBER"
            ]
          },
          "scheduled_at": {
            "type": "string",
            "format": "date-time",
            "title": "Scheduled At"
          },
          "doctor_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doctor Id"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "examples": [
              "scheduled"
            ]
          },
          "notified_via": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notified Via"
          }
        },
        "type": "object",
        "required": [
          "id",
          "patient_id",
          "tier",
          "scheduled_at",
          "status"
        ],
        "title": "AppointmentOut"
      },
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "examples": [
              "Not a recognized patient session"
            ]
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RouteAppointmentRequest": {
        "properties": {
          "conversation_id": {
            "type": "string",
            "format": "uuid",
            "title": "Conversation Id"
          }
        },
        "type": "object",
        "required": [
          "conversation_id"
        ],
        "title": "RouteAppointmentRequest"
      },
      "RouteAppointmentResponse": {
        "properties": {
          "appointment": {
            "$ref": "#/components/schemas/AppointmentOut"
          }
        },
        "type": "object",
        "required": [
          "appointment"
        ],
        "title": "RouteAppointmentResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Supabase session JWT",
        "scheme": "bearer"
      }
    }
  },
...
}